Форма авторизации вынесена в отдельный компонент
