+++ JS/api.js
@@ -1,4 +1,4 @@
-import { renderApp } from "./renderApp.js";
+import { renderApp, token } from "./renderApp.js";
 import { getCommentsList } from "./CommentsList.js";
 
 const listElem = document.getElementById('list-comments');
@@ -8,9 +8,6 @@ export let isInitialLoading = true;
 export let isPosting = false;
 
 const baseHost = "https://wedev-api.sky.pro/api/v2/freddy-krugliy/comments";
-
-export let token = null;
-// "Bearer asb4c4boc86gasb4c4boc86g37w3cc3bo3b83k4g37k3bk3cg3c03ck4k";
 
 
 function getFromApi(data) {
@@ -93,7 +90,6 @@ function postToApi(data, addCommentElem, addNameElem) {
         };
     })
 };
-
 
 
 export { getFromApi, postToApi };
